Move src files to app
